requirements.txt
click==8.1.3
importlib-metadata==5.0.0
Pillow==9.2.0
python-dateutil==2.8.2
six==1.16.0
tabulate==0.9.0
typing-extensions==4.4.0
zipp==3.9.0
